requirements.txt
streamlit==1.39.0
tensorflow
opencv-python
numpy
pandas
Pillow
